API jobs hsould run serially now.
The pre-start scripts for Cloud Controller kicks
off db migration tasks. If these are run in parallel,
they will fail on some nodes, since they all think the
database hasn't been migrate, but then someone will race
to create tables, and others will be unable to create tables.

When this was part of the launch script, monit would just restart
the job until it worked. Now that it's a pre-start script, BOSH
bails out on the first error. Serializing the deploys mitigates
the race conditions involved in db migration.